need to add one more test, page refresh
